+++ tool/release.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
-import 'package:yaml/yaml.dart';
+
 import 'package:path/path.dart' as path;
-import 'dart:convert';
+import 'package:yaml/yaml.dart';
 
 /// Automated release script for rdf_mapper
 ///